refactor(aletheia): extract runtime::setup submodule (#2642)
Move factory functions to a dedicated file: build_provider_registry,
build_tool_registry, create_embedding_provider, open_knowledge_store,
build_signal_provider, and start_inbound_dispatch. Trim mod.rs imports
to only what remains needed.
